+++ docs/CHANGELOG.rst
@@ -5,9 +5,9 @@ Change Log
 All notable changes to this project are documented in this file.
 
 
-==========
-Unreleased
-==========
+==================
+1.4.0 - 2016-10-19
+==================
 
 Added
 -----